Move all of the source code into a sources directory.
Gets rid of the .env file which was used to set the PYTHONPATH and adds a Makefil
which has rules to run each of the commands that we care about. Those make rules
set the PYTHONPATH for themselves. CI rules add the sources and tests directories.
Rules that run the app only add the sources directory.